refactor: split out the related article proptypes inline with other packages
