+++ packages/article/related-articles/related-article-item-proptypes.js
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import { treePropType } from "@times-components/markup";
 
-export const relatedArticleItemPropTypes = {
+export default {
   article: PropTypes.shape({
     byline: PropTypes.arrayOf(treePropType),
     headline: PropTypes.string.isRequired,
@@ -11,14 +11,4 @@ export const relatedArticleItemPropTypes = {
     url: PropTypes.string
   }).isRequired,
   onPress: PropTypes.func.isRequired
-};
-
-export const relatedArticlesPropTypes = {
-  articles: PropTypes.arrayOf(relatedArticleItemPropTypes.article),
-  onPress: PropTypes.func.isRequired,
-  template: PropTypes.string
-};
-
-export const defaultProps = {
-  template: "DEFAULT"
 };
